refactor: rename items game module to item
