+++ game/src/item/registry.rs
