+++ items/greek/src/bident_of_hades.rs
@@ -3,7 +3,7 @@ use {
         constants::*,
         prelude::*,
     },
-    mythmallow::items::constants::MELEE_ITEM_TAG,
+    mythmallow::item::constants::MELEE_ITEM_TAG,
 };
 
 /// Base price of the item.
