+++ items/greek/src/bow_of_artemis.rs
@@ -3,7 +3,7 @@ use {
         constants::*,
         prelude::*,
     },
-    mythmallow::items::constants::RANGED_ITEM_TAG,
+    mythmallow::item::constants::RANGED_ITEM_TAG,
 };
 
 /// Size of the item.
